+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # UNRELEASED
 
+- Changed: ***BREAKING*** `EntityEvent`s no longer bubble by default.
+  - If you are using `#[derive(EntityEvent)]`, you will need to add the `#[can_bubble]` attribute to
+    enable bubbling.
+  - If you are manually implementing the trait, you will need to override the default `can_bubble`
+    trait method and return true.
+- Changed: dissolved the `bevy_eventlistener_core` crate.
 - Changed: `commands_mut`, `target_commands_mut`, `target_component_mut`, `listener_commands_mut`,
   and `listener_component_mut` have been changed to provide a mutable reference to
   `ListenerInput<E>`. This now makes it possible to call `stop_propagation()` from these functions.